+++ TibiaAPI/Appearances/AppearanceStorage.cs
@@ -18,7 +18,7 @@ namespace OXGaming.TibiaAPI.Appearances
             lastOutfitId = appearances.Outfit.Aggregate((last, current) => last.Id > current.Id ? last : current).Id;
         }
 
-        public ObjectInstance CreateObjectInstance(ushort id, byte data)
+        public ObjectInstance CreateObjectInstance(uint id, uint data)
         {
             if (id >= (uint)CreatureInstanceType.Creature && id <= lastObjectId)
             {
@@ -27,7 +27,7 @@ namespace OXGaming.TibiaAPI.Appearances
             return null;
         }
 
-        public OutfitInstance CreateOutfitInstance(ushort id, byte colorHead, byte colorTorso, byte colorLegs, byte colorDetail, byte addons)
+        public OutfitInstance CreateOutfitInstance(uint id, byte colorHead, byte colorTorso, byte colorLegs, byte colorDetail, byte addons)
         {
             if (id >= 0 && id <= lastOutfitId)
             {
@@ -36,7 +36,7 @@ namespace OXGaming.TibiaAPI.Appearances
             return null;
         }
 
-        public Utilities.Appearance GetObjectType(ushort id)
+        public Utilities.Appearance GetObjectType(uint id)
         {
             if (id >= (uint)CreatureInstanceType.Creature && id <= lastObjectId)
             {
@@ -45,7 +45,7 @@ namespace OXGaming.TibiaAPI.Appearances
             return null;
         }
 
-        public Utilities.Appearance GetOutfitType(ushort id)
+        public Utilities.Appearance GetOutfitType(uint id)
         {
             if (id >= 1 && id <= lastOutfitId)
             {
